Blockless runtime path set as environment variable and proper closing of log files
